Include FOSJsBundle to use named routes
The FOSJSBundle provides a jsrouting-bundle that allows the routes to
be created in js file based on the backend settings. What this means
is that you do not have to hardcode the routes in angular based javascript
files but instead use the named routes from symfony.

Forked the original repo to add support for AngularJS. The new method called
generateAngularRoute takes a named route and creates a ui-routes based route
that could be used to render the screens.