Custom Form definitions to include angular models
Added a custom Twig snippet that will convert standard HTML Elements (for now
input, textarea and hidden) are converted into Angular Model based elements.
With this, just a FormType declaration should convert all these standard Symfony
form elements to use AngularJS model based definitions.

Added a directive(angular-submit) that will be added to each form(will need to
automate this through the above twig snippet). This directive adds a ngSubmit
directive to teh form element, which takes all elements of the form and sends
it through to the backend through $http. (Sending to backend not implemented in
this commit).

Click on the "Contact Form" to see this in action and the custom_form_div_layout.html.twig
for custom form declaration